Move UseCaseStore in store & untype store keys
