+++ frontend/src/domain/use_cases/error/displayOrLogError.ts
@@ -2,7 +2,7 @@ import { DisplayedError } from '../../../libs/DisplayedError'
 import { INITIAL_STATE, type DisplayedErrorState, displayedErrorActions } from '../../shared_slices/DisplayedError'
 import { setError } from '../../shared_slices/Global'
 
-import type { MainAppUseCase } from '../../../types'
+import type { MainAppUseCase } from '@store'
 
 /**
  * Dispatch:
