+++ frontend/src/features/SideWindow/useCases/confirmSideWindowDraftCancellationAndProceed.ts
@@ -1,13 +1,14 @@
-import { sideWindowActions } from '../../../domain/shared_slices/SideWindow'
-import { UseCaseStore, UseCaseStoreKey } from '../../../libs/UseCaseStore'
-import { assertNotNullish } from '../../../utils/assertNotNullish'
+import { UseCaseStore } from '@store/UseCaseStore'
+import { assertNotNullish } from '@utils/assertNotNullish'
 
-import type { MainAppThunk } from '../../../store'
+import { sideWindowActions } from '../../../domain/shared_slices/SideWindow'
+
+import type { MainAppThunk } from '@store'
 
 export const confirmSideWindowDraftCancellationAndProceed = (): MainAppThunk => (dispatch, getState) => {
   const { sideWindow } = getState()
   assertNotNullish(sideWindow.nextPath)
-  const pendingUseCase = UseCaseStore.get(UseCaseStoreKey.DRAFT_CANCELLATION_CONFIRMATION)
+  const pendingUseCase = UseCaseStore.get('MISSION_DRAFT_CANCELLATION_CONFIRMATION')
 
   if (pendingUseCase) {
     dispatch(pendingUseCase())
